requirements.txt
inference-gpu
pillow
gradio